requirements.txt
pyTelegramBotAPI==4.14.1
Flask==2.3.2
python-dotenv==1.0.1
